+++ pkg/controllers/applicationset_controller_test.go
@@ -13,9 +13,7 @@ import (
 	"github.com/argoproj/argo-cd/common"
 	"github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
 	argov1alpha1 "github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
-	"github.com/sirupsen/logrus"
 	log "github.com/sirupsen/logrus"
-	logtest "github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -1347,382 +1345,6 @@ func TestGetMinRequeueAfter(t *testing.T) {
 
 	assert.Equal(t, time.Duration(1)*time.Second, got)
 }
-
-func TestInvalidGenerators(t *testing.T) {
-
-	scheme := runtime.NewScheme()
-	err := argoprojiov1alpha1.AddToScheme(scheme)
-	assert.Nil(t, err)
-	err = argov1alpha1.AddToScheme(scheme)
-	assert.Nil(t, err)
-
-	for _, c := range []struct {
-		testName        string
-		appSet          argoprojiov1alpha1.ApplicationSet
-		expectedInvalid bool
-		expectedNames   map[string]bool
-	}{
-		{
-			testName: "valid generators, with annotation",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "name",
-					Namespace: "namespace",
-					Annotations: map[string]string{
-						"kubectl.kubernetes.io/last-applied-configuration": `{
-							"spec":{
-								"generators":[
-									{"list":{}},
-									{"cluster":{}},
-									{"git":{}}
-								]
-							}
-						}`,
-					},
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     &argoprojiov1alpha1.ListGenerator{},
-							Clusters: nil,
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: &argoprojiov1alpha1.ClusterGenerator{},
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      &argoprojiov1alpha1.GitGenerator{},
-						},
-					},
-				},
-			},
-			expectedInvalid: false,
-			expectedNames:   map[string]bool{},
-		},
-		{
-			testName: "invalid generators, no annotation",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "name",
-					Namespace: "namespace",
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-					},
-				},
-			},
-			expectedInvalid: true,
-			expectedNames:   map[string]bool{},
-		},
-		{
-			testName: "valid and invalid generators, no annotation",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "name",
-					Namespace: "namespace",
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     nil,
-							Clusters: &argoprojiov1alpha1.ClusterGenerator{},
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      &argoprojiov1alpha1.GitGenerator{},
-						},
-					},
-				},
-			},
-			expectedInvalid: true,
-			expectedNames:   map[string]bool{},
-		},
-		{
-			testName: "valid and invalid generators, with annotation",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "name",
-					Namespace: "namespace",
-					Annotations: map[string]string{
-						"kubectl.kubernetes.io/last-applied-configuration": `{
-							"spec":{
-								"generators":[
-									{"cluster":{}},
-									{"bbb":{}},
-									{"git":{}},
-									{"aaa":{}}
-								]
-							}
-						}`,
-					},
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     nil,
-							Clusters: &argoprojiov1alpha1.ClusterGenerator{},
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      &argoprojiov1alpha1.GitGenerator{},
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-					},
-				},
-			},
-			expectedInvalid: true,
-			expectedNames: map[string]bool{
-				"aaa": true,
-				"bbb": true,
-			},
-		},
-		{
-			testName: "invalid generator, annotation with missing spec",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "name",
-					Namespace: "namespace",
-					Annotations: map[string]string{
-						"kubectl.kubernetes.io/last-applied-configuration": `{
-						}`,
-					},
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-					},
-				},
-			},
-			expectedInvalid: true,
-			expectedNames:   map[string]bool{},
-		},
-		{
-			testName: "invalid generator, annotation with missing generators array",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "name",
-					Namespace: "namespace",
-					Annotations: map[string]string{
-						"kubectl.kubernetes.io/last-applied-configuration": `{
-							"spec":{
-							}
-						}`,
-					},
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-					},
-				},
-			},
-			expectedInvalid: true,
-			expectedNames:   map[string]bool{},
-		},
-		{
-			testName: "invalid generator, annotation with empty generators array",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "name",
-					Namespace: "namespace",
-					Annotations: map[string]string{
-						"kubectl.kubernetes.io/last-applied-configuration": `{
-							"spec":{
-								"generators":[
-								]
-							}
-						}`,
-					},
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-					},
-				},
-			},
-			expectedInvalid: true,
-			expectedNames:   map[string]bool{},
-		},
-		{
-			testName: "invalid generator, annotation with empty generator",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "name",
-					Namespace: "namespace",
-					Annotations: map[string]string{
-						"kubectl.kubernetes.io/last-applied-configuration": `{
-							"spec":{
-								"generators":[
-									{}
-								]
-							}
-						}`,
-					},
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-					},
-				},
-			},
-			expectedInvalid: true,
-			expectedNames:   map[string]bool{},
-		},
-	} {
-		hasInvalid, names := invalidGenerators(&c.appSet)
-		assert.Equal(t, c.expectedInvalid, hasInvalid, c.testName)
-		assert.Equal(t, c.expectedNames, names, c.testName)
-	}
-}
-
-func TestCheckInvalidGenerators(t *testing.T) {
-
-	scheme := runtime.NewScheme()
-	err := argoprojiov1alpha1.AddToScheme(scheme)
-	assert.Nil(t, err)
-	err = argov1alpha1.AddToScheme(scheme)
-	assert.Nil(t, err)
-
-	for _, c := range []struct {
-		testName    string
-		appSet      argoprojiov1alpha1.ApplicationSet
-		expectedMsg string
-	}{
-		{
-			testName: "invalid generator, without annotation",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-app-set",
-					Namespace: "namespace",
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     &argoprojiov1alpha1.ListGenerator{},
-							Clusters: nil,
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      &argoprojiov1alpha1.GitGenerator{},
-						},
-					},
-				},
-			},
-			expectedMsg: "ApplicationSet test-app-set contains unrecognized generators",
-		},
-		{
-			testName: "invalid generator, with annotation",
-			appSet: argoprojiov1alpha1.ApplicationSet{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-app-set",
-					Namespace: "namespace",
-					Annotations: map[string]string{
-						"kubectl.kubernetes.io/last-applied-configuration": `{
-							"spec":{
-								"generators":[
-									{"list":{}},
-									{"bbb":{}},
-									{"git":{}},
-									{"aaa":{}}
-								]
-							}
-						}`,
-					},
-				},
-				Spec: argoprojiov1alpha1.ApplicationSetSpec{
-					Generators: []argoprojiov1alpha1.ApplicationSetGenerator{
-						{
-							List:     &argoprojiov1alpha1.ListGenerator{},
-							Clusters: nil,
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      &argoprojiov1alpha1.GitGenerator{},
-						},
-						{
-							List:     nil,
-							Clusters: nil,
-							Git:      nil,
-						},
-					},
-				},
-			},
-			expectedMsg: "ApplicationSet test-app-set contains unrecognized generators: aaa, bbb",
-		},
-	} {
-		oldhooks := logrus.StandardLogger().ReplaceHooks(logrus.LevelHooks{})
-		defer logrus.StandardLogger().ReplaceHooks(oldhooks)
-		hook := logtest.NewGlobal()
-
-		checkInvalidGenerators(&c.appSet)
-		assert.True(t, len(hook.Entries) >= 1, c.testName)
-		assert.Equal(t, logrus.WarnLevel, hook.LastEntry().Level, c.testName)
-		assert.Equal(t, c.expectedMsg, hook.LastEntry().Message, c.testName)
-		hook.Reset()
-	}
-}
-
 func TestHasDuplicateNames(t *testing.T) {
 
 	scheme := runtime.NewScheme()
