+++ pkg/generators/generator_spec_processor.go
@@ -1,0 +1,82 @@
+package generators
+
+import (
+	"reflect"
+
+	argoprojiov1alpha1 "github.com/argoproj-labs/applicationset/api/v1alpha1"
+	"github.com/imdario/mergo"
+	log "github.com/sirupsen/logrus"
+)
+
+func GetRelevantGenerators(requestedGenerator *argoprojiov1alpha1.ApplicationSetGenerator, generators map[string]Generator) []Generator {
+	var res []Generator
+
+	v := reflect.Indirect(reflect.ValueOf(requestedGenerator))
+	for i := 0; i < v.NumField(); i++ {
+		field := v.Field(i)
+		if !field.CanInterface() {
+			continue
+		}
+
+		if !reflect.ValueOf(field.Interface()).IsNil() {
+			res = append(res, generators[v.Type().Field(i).Name])
+		}
+	}
+
+	return res
+}
+
+type TransformResult struct {
+	Params   []map[string]string
+	Template argoprojiov1alpha1.ApplicationSetTemplate
+}
+
+//Temp transform a spec generator to list of params and a template
+func Transform(requestedGenerator argoprojiov1alpha1.ApplicationSetGenerator, allGenerators map[string]Generator, baseTemplate argoprojiov1alpha1.ApplicationSetTemplate) ([]TransformResult, error) {
+	res := []TransformResult{}
+	var firstError error
+
+	generators := GetRelevantGenerators(&requestedGenerator, allGenerators)
+	for _, g := range generators {
+		// we call mergeGeneratorTemplate first because GenerateParams might be more costly so we want to fail fast if there is an error
+		mergedTemplate, err := mergeGeneratorTemplate(g, &requestedGenerator, baseTemplate)
+		if err != nil {
+			log.WithError(err).WithField("generator", g).
+				Error("error generating params")
+			if firstError == nil {
+				firstError = err
+			}
+			continue
+		}
+
+		params, err := g.GenerateParams(&requestedGenerator)
+		if err != nil {
+			log.WithError(err).WithField("generator", g).
+				Error("error generating params")
+			if firstError == nil {
+				firstError = err
+			}
+			continue
+		}
+
+		res = append(res, TransformResult{
+			Params:   params,
+			Template: mergedTemplate,
+		})
+
+	}
+
+	return res, firstError
+
+}
+
+func mergeGeneratorTemplate(g Generator, requestedGenerator *argoprojiov1alpha1.ApplicationSetGenerator, applicationSetTemplate argoprojiov1alpha1.ApplicationSetTemplate) (argoprojiov1alpha1.ApplicationSetTemplate, error) {
+
+	// Make a copy of the value from `GetTemplate()` before merge, rather than copying directly into
+	// the provided parameter (which will touch the original resource object returned by client-go)
+	dest := g.GetTemplate(requestedGenerator).DeepCopy()
+
+	err := mergo.Merge(dest, applicationSetTemplate)
+
+	return *dest, err
+}
