+++ pkg/utils/util.go
@@ -4,11 +4,15 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"reflect"
+	"sort"
 	"strconv"
 	"strings"
 
+	argoprojiov1alpha1 "github.com/argoproj-labs/applicationset/api/v1alpha1"
 	argov1alpha1 "github.com/argoproj/argo-cd/pkg/apis/application/v1alpha1"
 	"github.com/pkg/errors"
+	log "github.com/sirupsen/logrus"
 	"github.com/valyala/fasttemplate"
 )
 
@@ -21,7 +25,7 @@ type Render struct {
 
 func (r *Render) RenderTemplateParams(tmpl *argov1alpha1.Application, params map[string]string) (*argov1alpha1.Application, error) {
 	if tmpl == nil {
-		return nil, fmt.Errorf("Application template is empty ")
+		return nil, fmt.Errorf("application template is empty ")
 	}
 
 	if len(params) == 0 {
@@ -82,4 +86,88 @@ func (r *Render) replace(fstTmpl *fasttemplate.Template, replaceMap map[string]s
 	}
 
 	return replacedTmpl, nil
+}
+
+// Log a warning if there are unrecognized generators
+func CheckInvalidGenerators(applicationSetInfo *argoprojiov1alpha1.ApplicationSet) {
+	hasInvalidGenerators, invalidGenerators := invalidGenerators(applicationSetInfo)
+	if len(invalidGenerators) > 0 {
+		gnames := []string{}
+		for n := range invalidGenerators {
+			gnames = append(gnames, n)
+		}
+		sort.Strings(gnames)
+		aname := applicationSetInfo.ObjectMeta.Name
+		msg := "ApplicationSet %s contains unrecognized generators: %s"
+		log.Warnf(msg, aname, strings.Join(gnames, ", "))
+	} else if hasInvalidGenerators {
+		name := applicationSetInfo.ObjectMeta.Name
+		msg := "ApplicationSet %s contains unrecognized generators"
+		log.Warnf(msg, name)
+	}
+}
+
+// Return true if there are unknown generators specified in the application set.  If we can discover the names
+// of these generators, return the names as the keys in a map
+func invalidGenerators(applicationSetInfo *argoprojiov1alpha1.ApplicationSet) (bool, map[string]bool) {
+	names := make(map[string]bool)
+	hasInvalidGenerators := false
+	for index, generator := range applicationSetInfo.Spec.Generators {
+		v := reflect.Indirect(reflect.ValueOf(generator))
+		found := false
+		for i := 0; i < v.NumField(); i++ {
+			field := v.Field(i)
+			if !field.CanInterface() {
+				continue
+			}
+			if !reflect.ValueOf(field.Interface()).IsNil() {
+				found = true
+				break
+			}
+		}
+		if !found {
+			hasInvalidGenerators = true
+			addInvalidGeneratorNames(names, applicationSetInfo, index)
+		}
+	}
+	return hasInvalidGenerators, names
+}
+
+func addInvalidGeneratorNames(names map[string]bool, applicationSetInfo *argoprojiov1alpha1.ApplicationSet, index int) {
+	// The generator names are stored in the "kubectl.kubernetes.io/last-applied-configuration" annotation
+	config := applicationSetInfo.ObjectMeta.Annotations["kubectl.kubernetes.io/last-applied-configuration"]
+	var values map[string]interface{}
+	err := json.Unmarshal([]byte(config), &values)
+	if err != nil {
+		log.Warnf("couldn't unmarshal kubectl.kubernetes.io/last-applied-configuration: %+v", config)
+		return
+	}
+
+	spec, ok := values["spec"].(map[string]interface{})
+	if !ok {
+		log.Warn("coundn't get spec from kubectl.kubernetes.io/last-applied-configuration annotation")
+		return
+	}
+
+	generators, ok := spec["generators"].([]interface{})
+	if !ok {
+		log.Warn("coundn't get generators from kubectl.kubernetes.io/last-applied-configuration annotation")
+		return
+	}
+
+	if index >= len(generators) {
+		log.Warnf("index %d out of range %d for generator in kubectl.kubernetes.io/last-applied-configuration", index, len(generators))
+		return
+	}
+
+	generator, ok := generators[index].(map[string]interface{})
+	if !ok {
+		log.Warn("coundn't get generator from kubectl.kubernetes.io/last-applied-configuration annotation")
+		return
+	}
+
+	for key := range generator {
+		names[key] = true
+		break
+	}
 }
